chore: Update Dashboard route group project structure :crossed_fingers:
- Organized folder structure to Feature Folder hierarchy:
  - Moved specific components from global to route group components folder if shared, to leaf folder if dedicated.
  - Nested components into folders for future tests.
  - Renamed component files to PascalCase.
- Updated import paths to match new file structure.
- Renamed form components for consistency.
- Added generateMetadata function to Signup and Reader Preferences pages.

Link: T-12219

Close: T-12239
