+++ src/app/(website)/(dashboard)/reader-preferences/PreferencesForm/PreferencesForm.js
@@ -1,20 +1,19 @@
+/* eslint-disable no-console */
 /* eslint-disable react/no-array-index-key */
 /* eslint-disable react/jsx-no-bind */
 "use client";
-
-import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { useState } from "react";
-import { useRef } from "react";
+import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 
-import FormProvider, { useFormData } from "@/components/providers/form-context";
-import ReactTurnstile from "@/components/turnstile";
 import { GlowingButton, Input, Lead } from "@/components/ui";
 import BackButton, { Checkbox, Radio } from "@/components/ui/forms";
-import hoverStyles from "@/lib/hover";
-import { addUserToList, updateUser } from "@/lib/server/actions";
-import { cx } from "@/lib/utils";
+import { getUserInfo, updateUser } from "@/lib/server/actions";
+
+import FormProvider, {
+  useFormData,
+} from "../../components/FormContext/FormContext";
+
 const TopicsArray = [
   "Accounting/Finance",
   "Artificial Intelligence",
@@ -32,7 +31,7 @@ const TopicsArray = [
   "Other",
 ];
 
-export default function MultiStepForm() {
+export default function PreferencesForm() {
   const [formStep, setFormStep] = useState(1);
 
   const nextFormStep = () => setFormStep((currentStep) => currentStep + 1);
@@ -40,9 +39,9 @@ export default function MultiStepForm() {
   const prevFormStep = () => setFormStep((currentStep) => currentStep - 1);
 
   return (
-    <div className="">
+    <div>
       <FormProvider>
-        <div className="mx-auto max-w-xl">
+        <div className="mx-auto mb-48 mt-12 max-w-xl">
           <StepOne
             formStep={formStep}
             prevFormStep={prevFormStep}
@@ -83,48 +82,38 @@ function StepOne({ formStep, nextFormStep }) {
     register,
   } = useForm({ mode: "onTouched" });
 
-  const formRef = useRef(null);
-
   const onSubmit = async (values) => {
-    try {
-      const ref = await formRef.current;
+    setLoading(true);
+    const result = await getUserInfo(values.email);
 
-      if (ref) {
-        const formData = new FormData(formRef.current);
-        const token = formData.get("cf-turnstile-response");
-        const res = await fetch("/api/turnstile-verify", {
-          method: "POST",
-          body: JSON.stringify({ token }),
-          headers: {
-            "content-type": "application/json",
-          },
-        });
-        const tokenResponse = await res.json();
-        if (tokenResponse.data.success) {
-          setLoading(true);
-          setFormValues(values);
-          const result = await addUserToList(values.email);
-          if (result.error) {
-            throw new Error("Fail to send to iterable");
-          }
-          nextFormStep();
-        }
-      }
-    } catch (error) {
-      setLoading(false);
+    if (result.error) {
+      console.log("Error fetching data");
     }
+
+    const { email, dataFields } = result || {};
+
+    const data = {
+      email,
+      firstName: dataFields.firstName,
+      topic1: dataFields.topic1,
+      topic2: dataFields.topic2,
+      topic3: dataFields.topic3,
+      topic4: dataFields.topic4,
+    };
+    setFormValues(data);
+    nextFormStep();
+    setLoading(false);
   };
 
   return (
     <div className={formStep === 1 ? "block" : "hidden"}>
       <div className="">
         <Lead className="text-center text-gray-200">
-          What{"\u2018"}s your email address?
-          <sup className="text-pink">&nbsp;*</sup>
+          Enter your email address
         </Lead>
       </div>
       <div className="mt-5">
-        <form ref={formRef} noValidate onSubmit={handleSubmit(onSubmit)}>
+        <form noValidate onSubmit={handleSubmit(onSubmit)}>
           <label className="sr-only">Email Address</label>
           <Input
             className="w-full border-neutral-200/10 bg-slate-900 text-white placeholder:text-gray-600"
@@ -144,23 +133,16 @@ function StepOne({ formStep, nextFormStep }) {
               },
             }}
           />
-          <ReactTurnstile />
+
           <div className="mt-10 flex justify-center">
             <GlowingButton
-              ariaLabel="Go to next step"
               type="submit"
               autoWidth
-              disabled={loading}>
-              {loading ? "Hold on..." : "Next"}
+              disabled={loading}
+              ariaLabel={loading ? "Please wait..." : "Change my Preferences"}>
+              {loading ? "Just a sec..." : "Change my Preferences"}
             </GlowingButton>
           </div>
-          <p className="mt-6 text-center text-xs leading-6 text-gray-400">
-            We care about your{" "}
-            <Link href="/privacy" className={cx("font-bold", hoverStyles)}>
-              privacy
-            </Link>
-            .
-          </p>
         </form>
       </div>
     </div>
@@ -168,13 +150,18 @@ function StepOne({ formStep, nextFormStep }) {
 }
 // Enter Name
 function StepTwo({ formStep, nextFormStep }) {
-  const { setFormValues } = useFormData();
-  //   console.log(formStep);
+  const { setFormValues, data } = useFormData();
+  console.log(data);
   const {
     handleSubmit,
     formState: { errors },
     register,
+    setValue,
   } = useForm({ mode: "all" });
+
+  useEffect(() => {
+    setValue("firstName", data.firstName);
+  }, [data, setValue]);
 
   const onSubmit = (values) => {
     setFormValues(values);
@@ -183,9 +170,12 @@ function StepTwo({ formStep, nextFormStep }) {
 
   return (
     <div className={formStep === 2 ? "block" : "hidden"}>
+      <p className="mb-10 text-center text-xs uppercase leading-6 tracking-wider text-gray-400">
+        Change name | Click next to skip
+      </p>
       <div className="">
         <Lead className="text-center text-gray-200">
-          👋🏾 What should we call you?<sup className="text-pink">&nbsp;*</sup>
+          👋🏾 What should we call you?
         </Lead>
       </div>
       <div className="mt-5">
@@ -196,8 +186,10 @@ function StepTwo({ formStep, nextFormStep }) {
             name="firstName"
             type="text"
             required
+            // defaultValue={data.firstName}
             placeholder="Enter your first name"
             aria-label="What's your first name?."
+            // onChange={handleChange}
             register={register}
             errors={errors}
             validations={{
@@ -219,12 +211,17 @@ function StepTwo({ formStep, nextFormStep }) {
 }
 // Primary Topic
 function StepThree({ formStep, nextFormStep }) {
-  const { setFormValues } = useFormData();
+  const { setFormValues, data } = useFormData();
   const {
     handleSubmit,
     formState: { errors },
     register,
+    setValue,
   } = useForm({ mode: "all" });
+
+  useEffect(() => {
+    setValue("topic1", data.topic1);
+  }, [data, setValue]);
 
   const onSubmit = (values) => {
     setFormValues(values);
@@ -233,10 +230,12 @@ function StepThree({ formStep, nextFormStep }) {
 
   return (
     <div className={formStep === 3 ? "block" : "hidden"}>
+      <p className="mb-10 text-center text-xs uppercase leading-6 tracking-wider text-gray-400">
+        Change your Primary Topic
+      </p>
       <div className="">
         <Lead className="text-center text-gray-200">
           Tell us your primay topic of inteest, we{"\u2018"}ll filter the rest.
-          <sup className="text-pink">&nbsp;*</sup>
         </Lead>
       </div>
       <div className="mt-5">
@@ -268,7 +267,7 @@ function StepThree({ formStep, nextFormStep }) {
           )}
 
           <div className="mt-10 flex justify-center">
-            <GlowingButton ariaLabel="Go to next step" type="submit" autoWidth>
+            <GlowingButton type="submit" autoWidth>
               Next
             </GlowingButton>
           </div>
@@ -286,9 +285,14 @@ function StepFour({ formStep, prevFormStep, nextFormStep }) {
     formState: { errors },
     register,
     watch,
+    setValue,
   } = useForm({ mode: "all" });
 
   const topicsSelected = watch("topics");
+
+  useEffect(() => {
+    setValue("topics", [data.topic2, data.topic3, data.topic4]);
+  }, [data, setValue]);
 
   const onSubmit = (values) => {
     if (Array.isArray(values.topics)) {
@@ -314,6 +318,9 @@ function StepFour({ formStep, prevFormStep, nextFormStep }) {
 
   return (
     <div className={formStep === 4 ? "block" : "hidden"}>
+      <p className="mb-10 text-center text-xs uppercase leading-6 tracking-wider text-gray-400">
+        Change secondary topics
+      </p>
       <div className="">
         <Lead className="text-center text-gray-200">
           Awesome. Select three more topics you{"\u2018"}d like to keep tabs on.
@@ -349,11 +356,11 @@ function StepFour({ formStep, prevFormStep, nextFormStep }) {
           <div className="mt-10 flex justify-center">
             <BackButton onClick={prevFormStep} />
             <GlowingButton
+              type="submit"
+              autoWidth
               ariaLabel={
                 topicsSelected ? "Go to the next step" : "Skip this step"
-              }
-              type="submit"
-              autoWidth>
+              }>
               {topicsSelected ? "Next" : "Skip"}
             </GlowingButton>
           </div>
@@ -371,12 +378,11 @@ function FormSubmit({ formStep, nextFormStep, prevFormStep }) {
 
   const submitForm = async () => {
     setLoading(true);
-
     const { email, ...rest } = data;
     const result = await updateUser(email, rest);
 
     if (!result.success) {
-      throw new Error("Failed to submit the form");
+      console.error("Error fetching data");
     }
     router.push("/signup/confirm");
   };
@@ -385,11 +391,11 @@ function FormSubmit({ formStep, nextFormStep, prevFormStep }) {
     <div className={formStep === 5 ? "block" : "hidden"}>
       <div className="">
         <Lead className="text-center text-gray-200">
-          19 seconds, a new record {data.firstName ? data.firstName : ", w00t"}!
+          Ready to change your preferences,{" "}
+          {data.firstName ? data.firstName : "w00t"}!
         </Lead>
         <p className="mt-3 text-center text-gray-400">
-          Click submit, then please check your inbox to verify your email
-          address.
+          Click submit, and we will update your preferences.
         </p>
       </div>
 
@@ -399,8 +405,8 @@ function FormSubmit({ formStep, nextFormStep, prevFormStep }) {
           autoWidth
           onClick={submitForm}
           disabled={loading}
-          ariaLabel="Submit Form">
-          {(loading && "Just a sec...") || "Submit"}
+          aria-label="Submit form">
+          {(loading && "Just a sec...") || "Update my Preferences"}
         </GlowingButton>
       </div>
     </div>
